An alternative "stateless" Ziria semantics
The main idea is to make a distinction, on the one hand, between
the graph of a Ziria computation, which is just a static
description of the program's dataflow together with some
control-flow constraints, and the execution model, which is a
particular (dynamic) scheduling strategy for executing the graph.

This commit is a draft implementation of the above in which all nodes in the
dataflow graph run concurrently. I think it's possible to view the
previous tick-process execution model as a particular dynamic
scheduling policy applied to this "soup-of-nodes", but I haven't
yet implemented this.

While I use the term "dataflow graph", there is actually some
control flow hidden inside the implementation of `zbind', in the
form of a set of control-flow constraints: When we `zbind z1 f',
'f' must blocking read on the private control channel between `z1' and
itself before proceeding.

I believe the soup-of-nodes model has the same problems (wrt.
parallel pipe-in-bind, etc.) as the tick-process model. See the
comment near 'test_pipe_in_bind' for more discussion.  This
commit could serve as a starting point from which to explore possible
solutions.