Alternative #2 "stateless" Ziria semantics
See the comments in 'ZirPar.hs' for more details.  The main
difference from that implementation is that in this file, the
dataflow graph is unfolded independently of the scheduling policy
used to execute the graph, which makes it possible to play around
with multiple different schedules (but so far there's only one).